IndexGeneration class in a separate file. Cleanup.
