+++ utils/__init__.py
@@ -1,0 +1,3 @@
+import logging
+
+logging.basicConfig(format="%(levelname)s:%(message)s", level=logging.INFO)
